+++ projects/Mallard/src/components/Stars/Stars.tsx
