+++ projects/Mallard/src/components/front/items/small-items.tsx
@@ -3,8 +3,8 @@ import { StyleSheet } from 'react-native'
 import { ItemTappable, PropTypes } from './helpers/item-tappable'
 import { TextBlock } from './helpers/text-block'
 import { SuperHeroImageItem } from './super-items'
-import { Stars } from 'src/components/stars/stars'
-import { SportScore } from 'src/components/sportscore/sportscore'
+import { Stars } from 'src/components/Stars/Stars'
+import { SportScore } from 'src/components/SportScore/SportScore'
 import { View } from 'react-native'
 import { CAPIArticle } from 'src/common'
