+++ projects/Mallard/src/components/front/items/trail-image-view.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { StyleProp, StyleSheet, View } from 'react-native'
 import { CAPIArticle, ImageUse } from 'src/common'
-import { Stars } from 'src/components/stars/stars'
+import { Stars } from 'src/components/Stars/Stars'
 import { useMediaQuery } from 'src/hooks/use-screen'
 import { Breakpoints } from 'src/theme/breakpoints'
 import { ImageResource } from '../image-resource'
-import { SportScore } from 'src/components/sportscore/sportscore'
+import { SportScore } from 'src/components/SportScore/SportScore'
 import { ArticleType } from '../../../../../Apps/common/src'
 
 const trailImageViewStyles = StyleSheet.create({
