+++ projects/Mallard/src/components/sportscore/SportScore.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react-native'
 import { withKnobs, text } from '@storybook/addon-knobs'
-import { SportScore } from './sportscore'
+import { SportScore } from './SportScore'
 
 const sportScore = text('sportScore', 'Luton 3 - Watford 0')
