+++ projects/Mallard/src/components/stars/__tests__/Stars.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import TestRenderer, { ReactTestRendererJSON } from 'react-test-renderer'
-import { Stars } from '../stars'
+import { Stars } from '../Stars'
 
 const rating = 5
